Moved alarms logic to `ProcessEventsTask`
